requirements.txt
locust>=2.37.2,<3.0.0
PyYAML>=6.0.2,<7.0.0
playwright>=1.52.0,<2.0.0
rich>=14.0.0,<15.0.0
streamlit>=1.42.0,<2.0.0
pandas>=2.2.0,<3.0.0
plotly>=5.24.0,<6.0.0
